requirements.txt
matplotlib==3.3.3
numpy==1.19.2
torch==1.9.0+cu111
bchlib==0.14.0
six==1.15.0
pytorch_wavelets==1.3.0
opencv_python==4.5.2.54
torchvision==0.10.0+cu111
kornia==0.6.1
lpips==0.1.4
Pillow==8.4.0
tensorboardX==2.4.1
